requirements.txt
fire
PyYAML
#python-lxc (apt install python-lxc)